Moved Dot-generator file to output.dot.schemaspy
