+++ src/main/java/org/schemaspy/output/dot/schemaspy/DotConnectorFinder.java
@@ -17,7 +17,7 @@
  * License along with this library; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
  */
-package org.schemaspy.view;
+package org.schemaspy.output.dot.schemaspy;
 
 import org.schemaspy.model.Table;
 import org.schemaspy.model.TableColumn;
