+++ src/main/java/org/schemaspy/output/dot/schemaspy/DotFormatter.java
@@ -18,7 +18,7 @@
  * License along with this library; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
  */
-package org.schemaspy.view;
+package org.schemaspy.output.dot.schemaspy;
 
 import org.schemaspy.Config;
 import org.schemaspy.model.Database;
@@ -26,7 +26,9 @@ import org.schemaspy.model.ForeignKeyConstraint;
 import org.schemaspy.model.Table;
 import org.schemaspy.model.TableColumn;
 import org.schemaspy.output.dot.DotConfig;
-import org.schemaspy.view.DotNode.DotNodeConfig;
+import org.schemaspy.output.dot.schemaspy.DotNode.DotNodeConfig;
+import org.schemaspy.view.StyleSheet;
+import org.schemaspy.view.WriteStats;
 
 import java.io.PrintWriter;
 import java.util.*;
