+++ src/test/java/org/schemaspy/integrationtesting/SchemaMetaIT.java
@@ -35,8 +35,8 @@ import org.schemaspy.input.dbms.xml.SchemaMeta;
 import org.schemaspy.model.Database;
 import org.schemaspy.model.DbmsMeta;
 import org.schemaspy.model.ProgressListener;
+import org.schemaspy.output.dot.schemaspy.DotFormatter;
 import org.schemaspy.testing.H2MemoryRule;
-import org.schemaspy.view.DotFormatter;
 import org.schemaspy.view.WriteStats;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.CommandLineRunner;
