+++ src/test/java/org/schemaspy/output/html/mustache/diagrams/MustacheSummaryDiagramFactoryTest.java
@@ -10,7 +10,7 @@ import org.schemaspy.model.ImpliedForeignKeyConstraint;
 import org.schemaspy.model.ProgressListener;
 import org.schemaspy.model.Table;
 import org.schemaspy.output.diagram.DiagramException;
-import org.schemaspy.view.DotFormatter;
+import org.schemaspy.output.dot.schemaspy.DotFormatter;
 import org.schemaspy.view.MustacheTableDiagram;
 import org.schemaspy.view.WriteStats;
